requirements.txt
asgiref==3.7.2
colorama==0.4.6
crayons==0.4.0
Django==5.0.1
libmagic==1.0
numpy==1.26.2
opencv-python==4.9.0.80
piexif==1.1.3
pillow==10.2.0
python-magic-bin==0.4.14
sqlparse==0.4.4
stegano==0.11.3
stepic==0.5.0
tzdata==2023.4
